JB42 -> JB43:  Phone: Incoming Call in Background
Allows user to set that Incoming Call UI will remain in background
if screen is on. Instead, a notification is added to StatusBar, which
can be used to bring up the Incoming Call UI. Ringtone and vibration
still occur as per system settings.

This has the same effect as apps marked as "immersive". It prevents
the incoming call from backgrounding the current activity.

Change-Id: Id686d68f1156b6744244e51cf8ceef8dd651154d

Conflicts:
	src/com/android/phone/PhoneUtils.java